+++ plugins/content-manager/src/main/java/com/wazuh/contentmanager/ContentManagerPlugin.java
@@ -44,9 +44,10 @@ import java.util.function.Supplier;
 import com.wazuh.contentmanager.client.CommandManagerClient;
 import com.wazuh.contentmanager.index.ContentIndex;
 import com.wazuh.contentmanager.index.ContextIndex;
-import com.wazuh.contentmanager.model.commandmanager.Command;
+import com.wazuh.contentmanager.model.command.Command;
 import com.wazuh.contentmanager.settings.PluginSettings;
-import com.wazuh.contentmanager.utils.SnapshotHelper;
+import com.wazuh.contentmanager.utils.Privileged;
+import com.wazuh.contentmanager.utils.SnapshotManager;
 
 /** Main class of the Content Manager Plugin */
 public class ContentManagerPlugin extends Plugin
@@ -54,11 +55,12 @@ public class ContentManagerPlugin extends Plugin
     private static final Logger log = LogManager.getLogger(ContentManagerPlugin.class);
     private ContextIndex contextIndex;
     private ContentIndex contentIndex;
-    private Environment environment;
-    private ClusterService clusterService;
+    private SnapshotManager snapshotManager;
     private ThreadPool threadPool;
     private TransportService transportService;
     private CommandManagerClient commandManagerClient;
+    private ClusterService clusterService;
+    private Environment environment;
 
     @Override
     public Collection<Object> createComponents(
@@ -74,19 +76,24 @@ public class ContentManagerPlugin extends Plugin
             IndexNameExpressionResolver indexNameExpressionResolver,
             Supplier<RepositoriesService> repositoriesServiceSupplier) {
         PluginSettings.getInstance(environment.settings(), clusterService);
+        this.threadPool = threadPool;
+        this.clusterService = clusterService;
         this.contextIndex = new ContextIndex(client);
         this.contentIndex = new ContentIndex(client);
         this.environment = environment;
-        this.clusterService = clusterService;
         this.threadPool = threadPool;
         // Initialize the CommandManagerClient.
         this.commandManagerClient = CommandManagerClient.getInstance(client);
 
+        this.snapshotManager =
+                new SnapshotManager(environment, this.contextIndex, this.contentIndex, new Privileged());
         return Collections.emptyList();
     }
 
     /**
-     * Call the CTI API on startup and get the latest consumer information into an index
+     * The initialization requires the existence of the {@link ContentIndex#INDEX_NAME} index. For
+     * this reason, we use a ClusterStateListener to listen for the creation of this index by the
+     * "setup" plugin, to then proceed with the initialization.
      *
      * @param localNode local Node info
      */
@@ -95,11 +102,45 @@ public class ContentManagerPlugin extends Plugin
         log.info("Test mode: {}", PluginSettings.getInstance().getTest_mode());
         if (PluginSettings.getInstance().getTest_mode()) {
             log.info("Dev environment detected, posting test command.");
-            this.commandManagerClient.postCommand(Command.create("0"));
+            this.commandManagerClient.post(Command.create("0"));
         }
-        SnapshotHelper snapshotHelper =
-                new SnapshotHelper(this.threadPool, this.environment, this.contextIndex, this.contentIndex);
-        this.clusterService.addListener(snapshotHelper);
+        // Only cluster managers are responsible for the initialization.
+        if (localNode.isClusterManagerNode()) {
+            if (this.clusterService.state().routingTable().hasIndex(ContentIndex.INDEX_NAME)) {
+                this.start();
+            }
+
+            // To be removed once we include the Job Scheduler.
+            this.clusterService.addListener(
+                    event -> {
+                        if (event.indicesCreated().contains(ContentIndex.INDEX_NAME)) {
+                            this.start();
+                        }
+                    });
+        }
+    }
+
+    /**
+     * Initialize. The initialization consists of:
+     *
+     * <pre>
+     *     1. fetching the latest consumer's information from the CTI API.
+     *     2. initialize from a snapshot if the local consumer does not exist, or its offset is 0.
+     * </pre>
+     */
+    private void start() {
+        try {
+            this.threadPool
+                    .generic()
+                    .execute(
+                            () -> {
+                                this.contextIndex.createIndex();
+                                this.snapshotManager.initialize();
+                            });
+        } catch (Exception e) {
+            // Log or handle exception
+            log.error("Error initializing snapshot helper: {}", e.getMessage(), e);
+        }
     }
 
     @Override