+++ plugins/content-manager/src/main/java/com/wazuh/contentmanager/client/HttpClient.java
@@ -57,9 +57,11 @@ public class HttpClient {
 
     private static final int TIMEOUT = 10;
     private static final Object LOCK = new Object();
-    // Singleton instance of the HTTP client
+
+    /** Singleton instance of the HTTP client. */
     protected static CloseableHttpAsyncClient httpClient;
-    // Base URI for API requests
+
+    /** Base URI for API requests. */
     protected final URI apiUri;
 
     /**
@@ -68,6 +70,7 @@ public class HttpClient {
      * @param apiUri The base URI for API requests.
      */
     protected HttpClient(@NonNull URI apiUri) {
+        log.debug("Client initialized pointing at [{}]", apiUri);
         this.apiUri = apiUri;
         startHttpAsyncClient();
     }