+++ plugins/content-manager/src/main/java/com/wazuh/contentmanager/settings/PluginSettings.java
@@ -22,7 +22,8 @@ import org.opensearch.cluster.service.ClusterService;
 import org.opensearch.common.settings.Setting;
 import org.opensearch.common.settings.Settings;
 
-import com.wazuh.contentmanager.utils.ClusterInfoHelper;
+import com.wazuh.contentmanager.utils.ClusterInfo;
+
 import reactor.util.annotation.NonNull;
 
 /** Singleton class to manage the plugin's settings. */
@@ -31,9 +32,6 @@ public class PluginSettings {
 
     /** Singleton instance. */
     private static PluginSettings INSTANCE;
-
-    /** Content Manager Plugin API path. */
-    public static final String API_BASE_URI = "/_plugins/_content_manager";
 
     /** Base Wazuh CTI URL */
     public static final String CTI_URL = "https://cti.wazuh.com";
@@ -58,6 +56,9 @@ public class PluginSettings {
     public static final Setting<Boolean> TEST_MODE =
             Setting.boolSetting(
                     "wazuh-test-mode", false, Setting.Property.NodeScope, Setting.Property.Dynamic);
+    /** Timeout of indexing operations */
+    public static final long TIMEOUT = 10L;
+
     private final String ctiBaseUrl;
     private final ClusterService clusterService;
     private final Boolean test_mode;
@@ -72,7 +73,7 @@ public class PluginSettings {
         this.clusterService = clusterService;
         this.test_mode = TEST_MODE.get(settings);
 
-        log.debug("Settings.loaded: {}", this.toString());
+        log.debug("Settings loaded: {}", this.toString());
     }
 
     /**
@@ -119,7 +120,7 @@ public class PluginSettings {
      * @return a string with the Content Manager full URL
      */
     public String getClusterBaseUrl() {
-        return ClusterInfoHelper.getClusterBaseUrl(this.clusterService);
+        return ClusterInfo.getClusterBaseUrl(this.clusterService);
     }
 
     public Boolean getTest_mode() {